熟悉页播放次数是小数：整数列让每次提交都被 422 挡下
Yoe 反馈「一直显示待上传，我明明有网」。不是网络问题——后端日志里
321 个 422，上传全被拒，客户端队列无限重试。

422 默认不记原因，日志里只有一行状态码。先加了校验失败日志
（只记字段路径与错误类型，不记值——请求体里有标注内容），
立刻定位到 `body.familiarization_plays` 的 int_from_float。

前端记的是「播完的圈数 + 当前这遍的进度」，本来就是小数；models.py
自己的注释写着「只播 0.3 遍就上手的人要单独看」——0.3 遍存不进整数列。
是 schema 与列定义错了，前端没错。

- Attempt.familiarization_plays: Integer → Float，AttemptIn/AttemptOut 同
- 迁移 c4d9a1e5b207 把线上列改成 double precision
- 补两条测试：小数能提交、能原样存取

顺带按 Yoe 的要求去掉界面上的保存/上传状态显示——那分散注意力，
而标注者对它也无能为力。侧栏打勾即代表标完并记下，上传情况看管理后台；
只保留「本机存储写不进去」这一种提示，那种情况下继续标就是白标。

后端 151 条、前端 59 条测试全过。

+++ server/app/models.py
@@ -145,8 +145,12 @@ class Attempt(Base):
     dimension: Mapped[str] = mapped_column(Text, nullable=False)
     # 熟悉页真正播了几遍。既是流程记录，也是质检信号——
     # 只播 0.3 遍就上手的人，后续标注质量要单独看。
-    familiarization_plays: Mapped[int] = mapped_column(
-        Integer, nullable=False, default=0
+    #
+    # **是小数不是整数**：前端记的是「播完的圈数 + 当前这遍的进度」，
+    # 「0.3 遍」这个说法本身就要求小数。定成整数会让每次提交都被 422 挡下，
+    # 而客户端的重试队列把这一次失败放大成无限重试。
+    familiarization_plays: Mapped[float] = mapped_column(
+        Float, nullable=False, default=0.0
     )
     status: Mapped[str] = mapped_column(Text, nullable=False)
     task_snapshot: Mapped[dict] = mapped_column(JsonCol, nullable=False)
